Increase the input data size in CloseInflaterDeflaterTest
CloseInflaterDeflater test executes a code path which should be expected to run
when writing on closed Inflater and Deflater. The test is making implicit
assumption about how system installed zlib library that JDK used would populate
output buffer when JDK  makes JNI call to deflate input buffer. On certain host
where system installed zlib library takes advantage of hardware compression,
this implicit assumptions may not hold true. Specifically in
testDeflaterOutputStream, it was expecting that when close() or finish() is
called on Deflater class, it would need total of two call to zlib's deflate
function to populate the output buffer, and as the buffer is called, any
attempt to write should end up with NullPointerException [1]. In case of zlib
using hardware exploitation for compression, this would not end up in
NullPointerException but will throw IOException [2]. This commit increases the
input data size for the test to 1024 bytes which would result in call to
DeflaterOutputStream.write resulting in expected NullPointerException after
Deflater is closed.

[1](https://github.com/r30shah/openj9-openjdk-jdk/blob/13cb8e01daf6f487a4ca157da38e41966ab30033/src/java.base/share/classes/java/util/zip/Deflater.java#L895)
[2](https://github.com/r30shah/openj9-openjdk-jdk/blob/13cb8e01daf6f487a4ca157da38e41966ab30033/src/java.base/share/classes/java/util/zip/DeflaterOutputStream.java#L199-L201)

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>